feat: remove http layer and use a custom wit interface
