+++ examples/components/http-bundling/demo.js
@@ -1,0 +1,5 @@
+import { csvUtils } from "./dist/transpiled/http_bundling_s.js"
+
+const DATA = "title,score\nHello,5\nWorld,10"
+
+console.log(csvUtils.calculateScore(DATA))
