--- examples/components/http-bundling/local.wadm.yaml
@@ -1,51 +1,0 @@
----
-apiVersion: core.oam.dev/v1beta1
-kind: Application
-metadata:
-  name: typescript-http-bundling
-  annotations:
-    description: 'Demo of Typescript HTTP hello world server that uses typescript dependencies'
-    wasmcloud.dev/authors: wasmCloud team
-    wasmcloud.dev/source-url: https://github.com/wasmCloud/wasmCloud/blob/main/examples/typescript/components/http-hello-world/wadm.yaml
-    wasmcloud.dev/readme-md-url: https://github.com/wasmCloud/wasmCloud/blob/main/examples/typescript/components/http-hello-world/README.md
-    wasmcloud.dev/homepage: https://github.com/wasmCloud/wasmCloud/tree/main/examples/typescript/components/http-hello-world
-    wasmcloud.dev/categories: |
-      http,http-server,typescript,example
-spec:
-  components:
-    # (Component) A test component that returns a string for any given HTTP request
-    - name: http-component
-      type: component
-      properties:
-        image: file://./build/http_bundling_s.wasm
-      traits:
-        # Govern the spread/scheduling of the component
-        - type: spreadscaler
-          properties:
-            instances: 1
-
-    # Add a capability provider that enables HTTP access
-    - name: httpserver
-      type: capability
-      properties:
-        image: ghcr.io/wasmcloud/http-server:0.23.2
-      traits:
-        # Link the httpserver to the component, and configure the HTTP server
-        # to listen on port 8000 for incoming requests
-        #
-        # Since this HTTP server provider calls the `http-component` component, we establish
-        # a unidirectional link from this `httpserver` provider (the "source")
-        # to the `http-component` component (the "target"), so the server can invoke
-        # the component to handle a request.
-        - type: link
-          properties:
-            target:
-              name: http-component
-            namespace: wasi
-            package: http
-            interfaces: [incoming-handler]
-            source:
-              config:
-                - name: default-http
-                  properties:
-                    address: 0.0.0.0:8000
