+++ examples/components/http-bundling/package.json
@@ -1,23 +1,20 @@
 {
-  "name": "http-hello-world",
+  "name": "bundled-rsbuild",
   "version": "0.1.0",
-  "description": "A WebAssembly Component that responds to HTTP requests using the `wasi:http` WIT interface",
-  "main": "dist/http-hello-world.js",
+  "description": "A WebAssembly Component implements a custom interface and uses npm dependencies",
+  "type": "module",
+  "main": "demo.js",
   "scripts": {
     "generate:types": "rimraf generated/types && jco types wit/ -o generated/types",
     "build:tsc": "tsc",
     "build:bundle": "rslib build",
-    "build:js": "jco componentize -w wit -o dist/http-bundling.wasm dist/esm/http-bundling.mjs",
+    "build:js": "jco componentize -w wit -o dist/bundled-rsbuild.wasm dist/esm/bundled-rsbuild.js",
     "build:component": "wash build --sign-only --config-path wasmcloud.toml",
     "build": "npm run generate:types && npm run build:tsc && npm run build:bundle && npm run build:js && npm run build:component",
+    "transpile": "jco transpile build/bundled_rsbuild_s.wasm -o dist/transpiled/",
+    "demo": "node demo.js",
+    "start": "npm run install-and-build && npm run transpile && npm run demo",
     "install-and-build": "npm install && npm run build",
-    "component:start": "wash start component file://build/http_bundling_s.wasm --auction-timeout-ms 10000 --timeout-ms 10000 typescript-http-bundling",
-    "component:stop": "wash stop component typescript-http-bundling",
-    "wadm:start": "wash app deploy local.wadm.yaml",
-    "wadm:stop": "wash app delete typescript-http-bundling v0.0.1",
-    "start": "npm run build && npm run wadm:start",
-    "reload": "npm run build && npm run wadm:stop && npm run wadm:start",
-    "stop": "npm run wadm:stop",
     "test": "echo \"Error: no test specified\" && exit 0"
   },
   "keywords": [],
