+++ examples/components/http-bundling/wkg.lock
@@ -1,12 +1,4 @@
 # This file is automatically generated.
 # It is not intended for manual editing.
 version = 1
-
-[[packages]]
-name = "wasi:http"
-registry = "wasi.dev"
-
-[[packages.versions]]
-requirement = "=0.2.3"
-version = "0.2.3"
-digest = "sha256:e526c1586efc94cd148e33725139be05c4bb58ba20466d348282bd8dc3999f1d"
+packages = []
